# runner/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Livepeer AI Runner",
        "description": "An application to run AI pipelines",
        "version": "v0.1.2"
    },
    "servers": [
        {
            "url": "https://dream-gateway.livepeer.cloud",
            "description": "Livepeer Cloud Community Gateway"
        }
    ],
    "paths": {
...
        "/text-to-image": {
            "post": {
                "summary": "Text To Image",
                "operationId": "text_to_image",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TextToImageParams"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/image-to-image": {
            "post": {
                "summary": "Image To Image",
                "operationId": "image_to_image",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_image_to_image_image_to_image_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "APIError": {
                "properties": {
                    "msg": {
                        "type": "string",
                        "title": "Msg"
                    }
                },
                "type": "object",
                "required": [
                    "msg"
                ],
                "title": "APIError"
            },
...
            "Body_image_to_image_image_to_image_post": {
                "properties": {
                    "prompt": {
                        "type": "string",
                        "title": "Prompt",
                        "description": "Text prompt(s) to guide image generation."
                    },
                    "image": {
                        "type": "string",
                        "format": "binary",
                        "title": "Image",
                        "description": "Uploaded image to modify with the pipeline."
                    },
                    "model_id": {
                        "type": "string",
                        "title": "Model Id",
                        "description": "Hugging Face model ID used for image generation.",
                        "default": ""
                    },
                    "loras": {
                        "type": "string",
                        "title": "Loras",
                        "description": "A lora and weight to use for image generation.",
                        "default": ""
                    },
                    "strength": {
                        "type": "number",
                        "title": "Strength",
                        "description": "Degree of transformation applied to the reference image (0 to 1).",
                        "default": 0.8
                    },
                    "guidance_scale": {
                        "type": "number",
                        "title": "Guidance Scale",
                        "description": "Encourages model to generate images closely linked to the text prompt (higher values may reduce image quality).",
                        "default": 7.5
                    },
                    "image_guidance_scale": {
                        "type": "number",
                        "title": "Image Guidance Scale",
                        "description": "Degree to which the generated image is pushed towards the initial image.",
                        "default": 1.5
                    },
                    "negative_prompt": {
                        "type": "string",
                        "title": "Negative Prompt",
                        "description": "Text prompt(s) to guide what to exclude from image generation. Ignored if guidance_scale < 1.",
                        "default": ""
                    },
                    "safety_check": {
                        "type": "boolean",
                        "title": "Safety Check",
                        "description": "Perform a safety check to estimate if generated images could be offensive or harmful.",
                        "default": true
                    },
                    "seed": {
                        "type": "integer",
                        "title": "Seed",
                        "description": "Seed for random number generation."
                    },
                    "num_inference_steps": {
                        "type": "integer",
                        "title": "Num Inference Steps",
                        "description": "Number of denoising steps. More steps usually lead to higher quality images but slower inference. Modulated by strength.",
                        "default": 100
                    },
                    "num_images_per_prompt": {
                        "type": "integer",
                        "title": "Num Images Per Prompt",
                        "description": "Number of images to generate per prompt.",
                        "default": 1
                    }
                },
                "type": "object",
                "required": [
                    "prompt",
                    "image"
                ],
                "title": "Body_image_to_image_image_to_image_post"
            },
...
            "HTTPError": {
                "properties": {
                    "detail": {
                        "$ref": "#/components/schemas/APIError"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "HTTPError"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ImageResponse": {
                "properties": {
                    "images": {
                        "items": {
                            "$ref": "#/components/schemas/Media"
                        },
                        "type": "array",
                        "title": "Images"
                    }
                },
                "type": "object",
                "required": [
                    "images"
                ],
                "title": "ImageResponse"
            },
            "Media": {
                "properties": {
                    "url": {
                        "type": "string",
                        "title": "Url"
                    },
                    "seed": {
                        "type": "integer",
                        "title": "Seed"
                    },
                    "nsfw": {
                        "type": "boolean",
                        "title": "Nsfw"
                    }
                },
                "type": "object",
                "required": [
                    "url",
                    "seed",
                    "nsfw"
                ],
                "title": "Media"
            },
...
            "TextToImageParams": {
                "properties": {
                    "model_id": {
                        "type": "string",
                        "title": "Model Id",
                        "description": "Hugging Face model ID used for image generation.",
                        "default": ""
                    },
                    "loras": {
                        "type": "string",
                        "title": "Loras",
                        "description": "A lora and weight to use for image generation. example: { \"nerijs/pixel-art-xl\" : 1.2 } "
                    },
                    "prompt": {
                        "type": "string",
                        "title": "Prompt",
                        "description": "Text prompt(s) to guide image generation. Separate multiple prompts with '|' if supported by the model."
                    },
                    "height": {
                        "type": "integer",
                        "title": "Height",
                        "description": "The height in pixels of the generated image.",
                        "default": 576
                    },
                    "width": {
                        "type": "integer",
                        "title": "Width",
                        "description": "The width in pixels of the generated image.",
                        "default": 1024
                    },
                    "guidance_scale": {
                        "type": "number",
                        "title": "Guidance Scale",
                        "description": "Encourages model to generate images closely linked to the text prompt (higher values may reduce image quality).",
                        "default": 7.5
                    },
                    "negative_prompt": {
                        "type": "string",
                        "title": "Negative Prompt",
                        "description": "Text prompt(s) to guide what to exclude from image generation. Ignored if guidance_scale < 1.",
                        "default": ""
                    },
                    "safety_check": {
                        "type": "boolean",
                        "title": "Safety Check",
                        "description": "Perform a safety check to estimate if generated images could be offensive or harmful.",
                        "default": true
                    },
                    "seed": {
                        "type": "integer",
                        "title": "Seed",
                        "description": "Seed for random number generation."
                    },
                    "num_inference_steps": {
                        "type": "integer",
                        "title": "Num Inference Steps",
                        "description": "Number of denoising steps. More steps usually lead to higher quality images but slower inference. Modulated by strength.",
                        "default": 50
                    },
                    "num_images_per_prompt": {
                        "type": "integer",
                        "title": "Num Images Per Prompt",
                        "description": "Number of images to generate per prompt.",
                        "default": 1
                    }
                },
                "type": "object",
                "required": [
                    "loras",
                    "prompt"
                ],
                "title": "TextToImageParams"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}